+++ MA_Thema_7/app/src/main/java/de/fhswf/ma/thema7/game/gameobjects/Player.java
@@ -13,7 +13,7 @@ public class Player implements GameObject
     private Bitmap playerImage;
     private RectF dst;
     private Point position;
-    private int radius = 64 * Constants.SCALE;
+    private int radius = (int) (64 * Constants.SCALE);
 
     public Player(Bitmap playerImage, RectF dst)
     {
