+++ MA_Thema_7/app/src/main/java/de/fhswf/ma/thema7/game/gameobjects/Wall.java
@@ -12,8 +12,8 @@ public class Wall implements GameObject
     private Bitmap wallImage;
     private RectF dst1;
     private RectF dst2;
-    private int wallWidth = 1080 * Constants.SCALE;
-    private int wallHeight = 128 * Constants.SCALE;
+    private int wallWidth = (int) (1080 * Constants.SCALE);
+    private int wallHeight = (int) (128 * Constants.SCALE);
 
     /**
      * @param wallImage Image used for the walls
