+++ MA_Thema_7/app/src/main/java/de/fhswf/ma/thema7/util/Constants.java
@@ -7,7 +7,7 @@ public class Constants
     public static int SCREEN_WIDTH;
     public static int SCREEN_HEIGHT;
 
-    public static int SCALE;
+    public static double SCALE;
 
     public static Context CURRENT_CONTEXT;
